Use single queue processing mechanism for loop
The event loop thread should use a single mechanism for processing
things placed on a queue (e.g. disposed threads, locking and in the
future the dispatch queue).

Additional minor changes have been made. Such as renaming some of the
private variables so the names make more sense in context.